ref: finishing definition of aktualizr + move some types into public API
